diagnosticsReport {rstudio}	R Documentation

Write an RStudio Diagnostics Report

Description

Write a diagnostics report to assist in troubleshooting problems with RStudio. The report is written to ~/rstudio-diagnostics and contains information on the R session and environment as well as copies of the RStudio log files.

Usage

diagnosticsReport()

Note

The diagnosticsReport function was added in version 0.97.139 of RStudio.

Examples

## Not run: 
rstudio::diagnosticsReport()

## End(Not run)

[Package rstudio version 0.98.945 Index]